% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment_merge.R
\name{align_merge}
\alias{align_merge}
\alias{tol_merge}
\title{Alignment merge of 'features' via m/z and RT}
\usage{
align_merge(
  main,
  sub,
  id = ".features_id",
  mz.main = "mz",
  mz.sub = "mz",
  rt.main = "rt.min",
  rt.sub = "rt.min",
  mz.tol = 0.05,
  rt.tol = 0.3,
  mz.weight = 75,
  rt.weight = 25,
  unique = T
)

tol_merge(
  main,
  sub,
  main_col = "mz",
  sub_col = "mz",
  tol = 0.002,
  bin_size = 1
)
}
\description{
Merge two data.frame of 'features' according to m/z and RT.
The formula is the same in MZmine2:
Score = (1 - rt.difference / rt.tolerance) * rt.weight +
(1 - mz.defference / mz.tolerance) * mz.weight

\code{align_merge}: ...

\code{tol_merge}: ...
}
