% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_draw.R
\name{arrow}
\alias{arrow}
\alias{parrow}
\alias{garrow}
\alias{garrow_city}
\alias{garrow_snake}
\alias{garrow_city_args}
\alias{sagnage}
\alias{sagnage_shiny}
\alias{maparrow}
\title{draw arrow}
\usage{
parrow(n = 5, col, type = "dashed", lwd = u(1, line))

garrow(p1, p2, args_line, args_arrow, fun_line, fun_arrow, city)

garrow_city(p1, p2, up, left, shift, gp_line)

garrow_snake(p1, p2, color, lwd = u(1, line), cur = 1)

garrow_city_args(
  shift = u(2, line),
  axis = "x",
  mid = 0.5,
  args_line = list(ncp = 1, curvature = 1, square = T),
  grob_anno = NULL,
  vp_anno = NULL,
  rev = F
)

sagnage(
  grob,
  left = T,
  l_gpar,
  borderF = 1.5,
  front_len = 0.1,
  vp_shift = u(1.5, line),
  ...
)

sagnage_shiny(label, left, color)

maparrow(
  obj,
  data,
  pattern = list(),
  round_cur = 0.3,
  pos = list(r = list(x = 1), l = list(x = 0), t = list(y = 1), b = list(y = 0))
)
}
\description{
...

\code{parrow}: ...

\code{garrow}: ...

\code{garrow_city}: ...

\code{garrow_snake}: ...

\code{garrow_city_args}: ...

\code{sagnage}: ...

\code{sagnage_shiny}: ...

\code{maparrow}: ...
}
