% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_identification.R
\docType{data}
\name{format_table}
\alias{format_table}
\alias{rename_table}
\alias{.filter_format}
\alias{.arrange_format}
\alias{.distinct_format}
\alias{.mutate_format}
\alias{.select_format}
\alias{.export_name}
\title{Format table via dplyr::*}
\format{
An object of class \code{list} of length 1.

An object of class \code{list} of length 3.

An object of class \code{list} of length 1.

An object of class \code{list} of length 4.

An object of class \code{character} of length 14.

An object of class \code{character} of length 17.
}
\usage{
rename_table(data, export_name = .export_name)

format_table(
  data,
  filter = .filter_format,
  arrange = .arrange_format,
  distinct = .distinct_format,
  mutate = .mutate_format,
  select = .select_format,
  export_name = .export_name
)

.filter_format

.arrange_format

.distinct_format

.mutate_format

.select_format

.export_name
}
\arguments{
\item{data}{data.frame. From \code{features_annotation(mcn)}.}
}
\description{
Format the data.frame via: \code{dplyr::filter}, \code{dplyr::arrange},
\code{dplyr::distinct}, \code{dplyr::mutate}, \code{dplyr::select},
\code{dplyr::rename}.

\code{rename_table}: ...

\code{format_table}: ...

\code{.filter_format}: ...

\code{.arrange_format}: ...

\code{.distinct_format}: ...

\code{.mutate_format}: ...

\code{.select_format}: ...

\code{.export_name}: ...
}
\keyword{datasets}
