% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_draw.R
\name{frame}
\alias{frame}
\alias{layout_row}
\alias{frame_row}
\alias{layout_col}
\alias{frame_col}
\alias{frame_place}
\title{draw in grid frame}
\usage{
layout_row(weight)

frame_row(weight, data, if.ex)

layout_col(weight)

frame_col(weight, data, if.ex)

frame_place(weight, data, type, if.ex)
}
\description{
...

\code{layout_row}: ...

\code{frame_row}: ...

\code{layout_col}: ...

\code{frame_col}: ...

\code{frame_place}: ...
}
