% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_draw.R
\name{nebulae_as_grob}
\alias{nebulae_as_grob}
\title{Convert Nebulae as 'grob' object}
\usage{
nebulae_as_grob(x)
}
\arguments{
\item{x}{\linkS4class{mcnebula} object.}
}
\description{
\code{nebulae_as_grob}: This will convert Child-Nebulae
as a 'grob' object. See package \code{grid} about 'grob' object.
}
