% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_draw.R
\name{network}
\alias{network}
\alias{fast_layout}
\alias{random_graph}
\title{Quickly draw random network diagrams}
\usage{
fast_layout(edges, layout = "fr", nodes = NULL)

random_graph(ids, n = 5, e = 4, layout = "fr")
}
\description{
...

\code{fast_layout}: ...

\code{random_graph}: ...
}
