% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_enrichment.R
\name{pathway_enrichment}
\alias{pathway_enrichment}
\alias{init_fella}
\alias{load_fella}
\alias{enrich_fella}
\alias{graph_fella}
\alias{plotGraph_fella}
\alias{cid.to.kegg}
\title{Perform pathway enrichment via package of 'FELLA'}
\usage{
init_fella(dir, org = c("hsa", "mmu", "rno"), seed = 1, rebuild = F)

load_fella(dir)

enrich_fella(id.lst, data)

graph_fella(
  obj.lst,
  data,
  method = c("pagerank", "diffusion", "hypergeom"),
  threshold = 0.1
)

plotGraph_fella(
  graph,
  layout = "graphopt",
  seed = 1,
  shape = c(Input = 15, Others = 16),
  color = c(Pathway = "#E64B35FF", Module = "#E377C2", Enzyme = "#EFC000", Reaction =
    "#4DBBD5FF", Compound = "#00A087FF"),
  size = c(Pathway = 7, Module = 5, Enzyme = 6, Reaction = 5, Compound = 10)
)

cid.to.kegg(cids)
}
\description{
Pathway enrichment analysis was performed using KEGG ID
via package of 'FELLA'.
(Convert CID to KEGG ID using the 'MetaboAnalystR' package.
See \url{https://github.com/xia-lab/MetaboAnalystR} for installation.)

\code{init_fella}: ...

\code{load_fella}: ...

\code{enrich_fella}: ...

\code{graph_fella}: ...

\code{plotGraph_fella}: Draw the graph via
package of 'ggplot2'.

\code{cid.to.kegg}: ...
}
\seealso{
\code{\link[FELLA:data-funs]{FELLA::buildDataFromGraph()}}, \code{\link[FELLA:data-funs]{FELLA::buildGraphFromKEGGREST()}}
}
