% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pick_annotation.R
\docType{data}
\name{pick_annotation}
\alias{pick_annotation}
\alias{pick_class}
\alias{.filter_pick.class}
\alias{PickClass}
\alias{pick_synonym}
\alias{.filter_pick.general}
\alias{PickGeneral}
\title{Pick unique annotation for compounds}
\format{
An object of class \code{list} of length 2.

An object of class \code{list} of length 6.
}
\usage{
pick_class(
  inchikey2d,
  class.rdata,
  filter = .filter_pick.class,
  fun = PickClass
)

.filter_pick.class

PickClass(class)

pick_synonym(
  inchikey2d = NULL,
  inchikey.rdata = NULL,
  synonym.rdata,
  iupac.rdata = NULL,
  filter = .filter_pick.general,
  fun = PickGeneral
)

.filter_pick.general

PickGeneral(
  syno,
  ps = c("^[a-zA-Z]*$", "^[a-zA-Z-]*$", "^[a-zA-Z0-9-]*$", "^[^:]*$")
)
}
\description{
Pick unique chemical class or synonyms for 'features'.

\code{pick_class}: ...

\code{.filter_pick.class}: ...

\code{PickClass}: ...

\code{pick_synonym}: ...

\code{.filter_pick.general}: ...

\code{PickGeneral}: ...
}
\seealso{
Other queries: 
\code{\link{query_classification}()},
\code{\link{query_inchikey}()},
\code{\link{query_iupac}()},
\code{\link{query_synonyms}()}
}
\concept{queries}
\keyword{datasets}
