% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot_heatmap.R
\name{plot_heatmap}
\alias{plot_heatmap}
\alias{handling_na}
\alias{log_trans}
\alias{add_tree.heatmap}
\alias{add_xgroup.heatmap}
\alias{add_xgroup.tile.heatmap}
\alias{add_ygroup.tile.heatmap}
\title{Plot heat map with ggplot2}
\usage{
plot_heatmap(
  id.lst,
  data,
  metadata,
  pal_class = (ggsci::pal_futurama())(12),
  pal_group,
  clust_row = T,
  clust_col = T,
  method = "complete"
)

handling_na(
  data,
  id.cols = c(".features_id"),
  metadata,
  sample.col = "sample",
  group.col = "group"
)

log_trans(
  data,
  id.cols = c(".features_id"),
  key = "sample",
  value = "value",
  set_min = T,
  factor = 10,
  fun = log2,
  center = T
)

add_tree.heatmap(df, p, clust_row = T, clust_col = T, method = "complete")

add_xgroup.heatmap(df, p)

add_xgroup.tile.heatmap(df, p, pal = NA)

add_ygroup.tile.heatmap(df, p, pal = NA)
}
\description{
According to list of 'ID' to draw mutiple heatmap...

\code{plot_heatmap}: ...

\code{handling_na}:
For each subset of data, the missing values will be filled with the average
value; if the set is all missing values, they will be filled with zero.

\code{log_trans}:
Convert wide data to long data; log transform the values; if there is a
value 0, replace it with 1/10 of the minimum value of the value column.

\code{add_tree.heatmap}: ...

\code{add_xgroup.heatmap}: ...

\code{add_xgroup.tile.heatmap}: ...

\code{add_ygroup.tile.heatmap}: ...
}
