% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_classification.R
\name{query_classification}
\alias{query_classification}
\alias{classyfire_get_classification}
\title{Query classification of compounds via packgage of 'classifyerR'}
\usage{
query_classification(
  inchikey2d,
  dir,
  inchikey.rdata = paste0(dir, "/inchikey.rdata"),
  rdata.name = "classification.rdata",
  classyfire_cl = NULL,
  planB = FALSE,
  gather_as_rdata = T,
  ...
)

classyfire_get_classification(
  sets,
  dir,
  classyfire_cl = NULL,
  log_file = paste0(dir, "/classyfire.log"),
  planB = FALSE,
  ...
)
}
\description{
The used function is:
\code{classyfireR::get_classification(inchikey)}

\code{query_classification}: ...

\code{classyfire_get_classification}: ...
}
\seealso{
Other queries: 
\code{\link{packing_as_rdata_list}()},
\code{\link{pick_annotation}},
\code{\link{query_fingerprints}()},
\code{\link{query_inchikey}()},
\code{\link{query_iupac}()}
}
\concept{queries}
