% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_fingerprints.R
\name{query_fingerprints}
\alias{query_fingerprints}
\alias{pubchem_get_fingerprints}
\alias{sdf_convert_tanimoto}
\title{Convert PubChem CID to tanimoto based chemical similarity}
\usage{
query_fingerprints(
  cid,
  dir,
  rdata.name = "fingerprints.rdata",
  curl_cl = NULL,
  extra = NULL,
  gather_as_rdata = T,
  group_number = 1,
  ...
)

pubchem_get_fingerprints(cid, dir, ...)

sdf_convert_tanimoto(sdf_list = NULL, dir, rdata.name = "fingerprints.rdata")
}
\description{
bulk search for tanimoto via pubchem API
(http://pubchem.ncbi.nlm.nih.gov/rest/pug/compound/cid/.../SDF)

\code{query_fingerprints}: ...

\code{pubchem_get_fingerprints}: ...

\code{sdf_convert_tanimoto}: ...
}
\seealso{
Other queries: 
\code{\link{packing_as_rdata_list}()},
\code{\link{pick_annotation}},
\code{\link{query_classification}()},
\code{\link{query_inchikey}()},
\code{\link{query_iupac}()}
}
\concept{queries}
