% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_inchikey.R
\name{query_inchikey}
\alias{query_inchikey}
\alias{pubchem_get_inchikey}
\title{Query InChIkey of compounds via 'InChIkey 2D'}
\usage{
query_inchikey(
  inchikey2d,
  dir,
  rdata.name = "inchikey.rdata",
  curl_cl = NULL,
  gather_as_rdata = T,
  ...
)

pubchem_get_inchikey(inchikey2d, dir, type = "inchikey", get = "InChIkey", ...)
}
\description{
The API:
url_start = paste0("https://pubchem.ncbi.nlm.nih.gov/rest/pug/compound/", type, "/")
url_end = paste0("/property/", paste(get, collapse = ","), "/CSV")
url = paste0(url_start, "/", inchikey2d, "/", url_end)

\code{query_inchikey}: ...

\code{pubchem_get_inchikey}: ...
}
\seealso{
Other queries: 
\code{\link{packing_as_rdata_list}()},
\code{\link{pick_annotation}},
\code{\link{query_classification}()},
\code{\link{query_fingerprints}()},
\code{\link{query_iupac}()}
}
\concept{queries}
