% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_others.R
\name{query_iupac}
\alias{query_iupac}
\title{Query IUPAC name of compounds via 'InChIkey 2D'}
\usage{
query_iupac(
  inchikey2d,
  dir,
  rdata.name = "iupac.rdata",
  curl_cl = NULL,
  gather_as_rdata = T,
  ...
)
}
\description{
Similar to \code{\link[=query_inchikey]{query_inchikey()}}, but get 'IUPACName'.

\code{query_iupac}: ...
}
\seealso{
Other queries: 
\code{\link{packing_as_rdata_list}()},
\code{\link{pick_annotation}},
\code{\link{query_classification}()},
\code{\link{query_fingerprints}()},
\code{\link{query_inchikey}()}
}
\concept{queries}
