% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_fingerprints.R, R/query_synonyms.R
\name{packing_as_rdata_list}
\alias{packing_as_rdata_list}
\alias{query_synonyms}
\alias{pubchem_get_synonyms}
\alias{grouping_vec2list}
\alias{extract_rdata_list}
\title{Query synonyms of compounds via CID}
\usage{
packing_as_rdata_list(
  path,
  pattern,
  rdata,
  extra = NULL,
  rm_files = T,
  dedup = T
)

query_synonyms(
  cid,
  dir,
  rdata.name = "synonyms.rdata",
  curl_cl = NULL,
  gather_as_rdata = T,
  group_number = 50,
  ...
)

pubchem_get_synonyms(cid, dir, ...)

grouping_vec2list(vector, group_number, byrow = F)

extract_rdata_list(rdata, names = NA)
}
\description{
\code{packing_as_rdata_list}: gather table as .rdata

Bulk search for compound synonyms via pubchem API.
(https://pubchem.ncbi.nlm.nih.gov/rest/pug/compound/cid/.../synonyms/XML)

\code{query_synonyms}: ...

\code{pubchem_get_synonyms}: ...

\code{grouping_vec2list}: ...

\code{extract_rdata_list}: extract results from .rdata
}
\seealso{
Other queries: 
\code{\link{pick_annotation}},
\code{\link{query_classification}()},
\code{\link{query_fingerprints}()},
\code{\link{query_inchikey}()},
\code{\link{query_iupac}()}
}
\concept{queries}
