% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_draw.R
\name{rect}
\alias{rect}
\alias{grect}
\alias{grecti}
\alias{grecti2}
\alias{grecti3}
\alias{grecto}
\alias{grectn}
\alias{grectn_frame}
\alias{lst_grecti}
\alias{grectN}
\alias{grecta}
\alias{gshiny}
\alias{glayer}
\title{a mutate of grid::rectGrob}
\usage{
grect(
  name,
  tfill = "#E18727FF",
  bfill = "white",
  t_args,
  tgp_args,
  t = roundrectGrob,
  b_args,
  bgp_args,
  b = roundrectGrob,
  order = c(1, 2),
  vp = NULL
)

grecti(
  label,
  cex = 1,
  x = 0.5,
  y = 1.005,
  borderF = 2,
  just = c("center", "top"),
  tfill = "#E18727FF",
  vp = .grecti.vp,
  order = c(2, 1),
  cvp_clip = "on",
  cvp_fix = T,
  ...
)

grecti2(label, cex = 1, tfill = "#E18727FF", borderF = list(2), ...)

grecti3(label, cex = 1, tfill = "#E18727FF", borderF = list(2), ...)

grecto(
  label,
  cex = 1,
  x = 0.5,
  y = 0.995,
  borderF = 2,
  just = c("center", "bottom"),
  tfill = "#E18727FF",
  vp = .grecto.vp,
  order = c(1, 2),
  cvp_clip = "on",
  cvp_fix = F,
  ...
)

grectn(
  bfill = "white",
  b_args,
  bgp_args,
  b = roundrectGrob,
  cvp_clip = "inherit"
)

grectn_frame(content, title, zo = T)

lst_grecti(names, pal, tar = "slot", fun = grecti, ...)

grectN(lab.1, lab.2, gp = gpar(fontface = "plain"), bfill = "white")

grecta(label, cex = 4)

gshiny(
  xn = 4,
  yn = 3,
  xps = seq(0, 1, , xn),
  yps = seq(0, 1, , yn),
  size = c(0.15, 0.02),
  color = .default_color,
  rect = rectGrob(),
  vp = viewport(clip = "off"),
  cvp = viewport(, , 0.9, 0.9)
)

glayer(n = 5, to = 0.2, gp = gpar(fill = "white"), fun = rectGrob)
}
\description{
...

\code{grect}: ...

\code{grecti}: ...

\code{grecti2}: ...

\code{grecti3}: ...

\code{grecto}: ...

\code{grectn}: ...

\code{grectn_frame}: ...

\code{lst_grecti}: ...

\code{grectN}: ...

\code{grecta}: ...

\code{gshiny}: ...

\code{glayer}: ...
}
