% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_select.R
\name{select_features}
\alias{select_features}
\alias{cross_select}
\title{Select 'features' for MCnebula2}
\usage{
select_features(
  mcn,
  classes = unique(nebula_index(mcn)$class.name),
  q.value = 0.05,
  logfc = 0.3,
  coef = NULL,
  tani.score_cutoff = NULL,
  order_by_coef = NULL,
  togather = F
)

cross_select(data.lst, filter.lst, target, split = NULL)
}
\description{
Select significant 'features' from MCnebula2 with
statistic results for downstream analysis of metabolomics.

\code{select_features}: ...

\code{cross_select}: ...
}
