% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_draw.R
\name{setnull}
\alias{setnull}
\alias{setnullvp}
\alias{grepPath}
\alias{sort_vpPaths}
\title{Set markers crossover viewports}
\usage{
setnull(target, args, name = "null")

setnullvp(pattern, args, x, name = NULL, fix = T, perl = F)

grepPath(pattern, x = NULL, grobs = T, viewports = T, perl = F)

sort_vpPaths(vpPaths)
}
\description{
...

\code{setnull}: ...

\code{setnullvp}: ...

\code{grepPath}: ...

\code{sort_vpPaths}: ...
}
