% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_draw.R
\name{text}
\alias{text}
\alias{gtext}
\alias{gtextp}
\alias{gtext0}
\alias{form}
\alias{gltext}
\title{a mutate of grid::textGrob}
\usage{
gtext(label, gp_arg, form = T, ...)

gtextp(label, gp_arg, form = T, ...)

gtext0(label, gp_arg, form = T, ...)

form(label)

gltext(
  label,
  gp_arg = list(),
  args = list(),
  l_gp = .gpar_dotted_line,
  flip = F,
  borderF = 1.2
)
}
\description{
...

\code{gtext}: ...

\code{gtextp}: ...

\code{gtext0}: ...

\code{form}: ...

\code{gltext}: ...
}
