% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\docType{data}
\name{utilites}
\alias{utilites}
\alias{tmp_pdf}
\alias{op}
\alias{.cairosvg_to_grob}
\alias{.as_dic}
\alias{fill_list}
\alias{n}
\alias{namel}
\alias{repSuffix}
\alias{\%>\%}
\alias{\%<>\%}
\alias{.expath}
\alias{agroup}
\alias{write_tsv}
\alias{read_tsv}
\alias{mapply_rename_col}
\alias{turn_vector}
\alias{group_switch}
\alias{.find_and_sort_strings}
\alias{maps}
\alias{order_list}
\alias{molconvert_structure}
\alias{obj.size}
\title{utilites for programming}
\format{
An object of class \code{character} of length 1.
}
\usage{
tmp_pdf()

op(file)

.cairosvg_to_grob(path)

.as_dic(vec, names, default, fill = T, as.list = T, na.rm = F)

fill_list(names, vec, default = vec[1])

n(name, n)

namel(...)

repSuffix(chs, anno = ".rep.")

lhs \%>\% rhs

lhs \%<>\% rhs

.expath

agroup(group, value, FUN.VALUE = character(1))

write_tsv(x, filename, col.names = T, row.names = F)

read_tsv(path)

mapply_rename_col(mutate_set, replace_set, names, fixed = F)

turn_vector(vec)

group_switch(data, meta.lst, by)

.find_and_sort_strings(strings, patterns)

maps(data, value, from, to)

order_list(list)

molconvert_structure(smile, path)

obj.size(x, ...)
}
\description{
This is a combination of tools that are not always used.

\code{tmp_pdf}: ...

\code{op}: ...

\code{.cairosvg_to_grob}: Convert cairo svg to 'grob'.

\code{.as_dic}: ...

\code{fill_list}: ...

\code{n}: ...

\code{namel}: ...

\code{repSuffix}: ...

\code{\%>\%}: ...

\code{\%<>\%}: ...

\code{.expath}: ...

\code{agroup}: ...

\code{write_tsv}: ...

\code{read_tsv}: ...

\code{mapply_rename_col}: ...

\code{turn_vector}: ...

\code{group_switch}: ...

\code{.find_and_sort_strings}: ...

\code{maps}: ...

\code{order_list}: ...

\code{molconvert_structure}: ...

\code{obj.size}: ...
}
\keyword{datasets}
