% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_draw.R
\name{zoom_pdf}
\alias{zoom_pdf}
\title{Zoom in locally pdf to png}
\usage{
zoom_pdf(
  file,
  position = c(0.5, 0.5),
  size = c(0.15, 0.1),
  page = 1,
  dpi = 2000,
  as.grob = T
)
}
\description{
...

\code{zoom_pdf}: ...
}
